Add files via upload
This is the project as of right now.

This is made on godot version 4.2

Note:
This does have *errors* when changing rooms and the continue button is disabled until i can figure out how to make game saves and access those saves. i

if there is an easier way to do ANY of the things i have done or messed up on, please feel free to change anything and let me know!